yanglint: schema-mount extension callback
Add a commandline argument ("-S") that accepts a file containing a schema
context to be used by the schema-mount extension.  This file has the same
format as what is provided to the "-Y" option.  A callback function for
the schema-mount extenion is registered if the new option is specified.

This allows validating instance data for models that include a schema
mount point.